Change database structure and add more services
Add repositories for: customer, activityReportTemplate

+++ JobTracker.backend/src/main/java/net/greenbeansit/jobtracker/server/data/repository/ActivityReportEntityRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import net.greenbeansit.jobtracker.server.data.entity.ActivityReportEntity;
 
-public interface ActivityReportEntityRepository extends CrudRepository<ActivityReportEntity, Long>
+public interface ActivityReportEntityRepository extends CrudRepository<ActivityReportEntity, Integer>
 {
 
 	void delete(ActivityReportEntity deleted);
@@ -17,8 +17,6 @@ public interface ActivityReportEntityRepository extends CrudRepository<ActivityR
 
 	@SuppressWarnings("unchecked")
 	ActivityReportEntity save(ActivityReportEntity persisted);
-
-	//ActivityReportEntity update(ActivityReportEntity persisted);
 
 	List<ActivityReportEntity> findByAuthor(Integer author);
